pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "setuptools-scm>=8.0"]
build-backend = "setuptools.build_meta"

[project]
name = "simple-astro-cap"
version = "0.1.0"
description = "Simple keyboard-centric astronomy camera capture tool"
requires-python = ">=3.11"
dependencies = [
    "PySide6>=6.6",
    "numpy>=1.26",
    "Pillow>=10.0",
]

[project.optional-dependencies]
dev = ["pytest>=7.0", "pytest-qt>=4.0"]

[project.scripts]
simple-astro-cap = "simple_astro_cap.__main__:main"

[project.gui-scripts]
simple-astro-cap-gui = "simple_astro_cap.__main__:main"

[tool.setuptools.packages.find]
where = ["src"]
